% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_temp.R
\name{fit_neutral_temp}
\alias{fit_neutral_temp}
\title{Estimate Turbulent Sensible Heat Flux}
\usage{
fit_neutral_temp(
  ref_temp,
  surf_temp,
  rho_air,
  zHeight,
  u_star,
  z_0m,
  k = 0.4,
  C_p = 1004,
  d_0 = NA
)
}
\arguments{
\item{rho_air}{density of moist air}

\item{zHeight}{height above the surface (metres)}

\item{u_star}{friction velocity}

\item{z_0m}{roughness length}

\item{k}{0.4 (von Karman’s constant)}

\item{C_p}{1004 (specific heat at constant pressure J kg-1 K-1}

\item{d_0}{displacement height due to veg (m)}

\item{tDiffMeas}{temp - temp at surface (kelvin)}
}
\value{

}
\description{
Estimate Turbulent Sensible Heat Flux
}
