% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_wind.R
\name{fit_neutral_wind}
\alias{fit_neutral_wind}
\title{Calculate Ustar, Z_0m, and, d_0 Given Measured Wind Speed}
\usage{
fit_neutral_wind(uMeas, zHeight, d_0 = NA)
}
\arguments{
\item{uMeas}{measured wind speed (m/s)}

\item{zHeight}{height of wind measurement (m)}
}
\value{
List of optimization result. Params are accessed by 'returnedobj'$par_est.
}
\description{
Calculate Ustar, Z_0m, and, d_0 Given Measured Wind Speed
}
\examples{
df <- data.frame(
U_1 = c(4.6, 6.0, 7.6, 9.0),

Z_1 = c(1, 3, 10, 30)
)



# Find parameters
fitted_wind <- fit_neutral_wind(df$U_1, df$Z_1)

# Put estimated parameters in data frame
est_pars <- data.frame(
 ustar = fitted_wind$par_est$par[1],

 z_0m = fitted_wind$par_est$par[2],

 d_0 = fitted_wind$par_est$par[3]
)
Use helper function fit_neutral_wind_helpr to confirm worked




}
