% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_wind.R
\name{fit_neutral_wind_helpr}
\alias{fit_neutral_wind_helpr}
\title{Helper Function to Estimate Wind Speed}
\usage{
fit_neutral_wind_helpr(params, zHeight)
}
\arguments{
\item{params}{estimated from fit_neutral_wind function (list of 3 params)}

\item{zHeight}{heights that you want wind speed estimated at}
}
\value{
list of estimated wind speeds
}
\description{
Helper Function to Estimate Wind Speed
}
\examples{
 fit_neutral_wind_helpr(c(est_pars$ustar, est_pars$z_0m, est_pars$d_0), meas_wind$height_m)
}
