% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monin_obukhov_sim.R
\name{monin_obukhov_sim}
\alias{monin_obukhov_sim}
\title{Monin Obukov Simularity Variable}
\usage{
monin_obukhov_sim(
  u_star,
  T_c,
  rho_air,
  z,
  Q_h,
  d_0,
  g = 9.81,
  k = 0.4,
  C_p = 1004.67
)
}
\arguments{
\item{u_star}{friction velocity (m/s)}

\item{T_c}{air temperature (celsuis)}

\item{rho_air}{density of moist air}

\item{z}{measurement height (m)}

\item{Q_h}{surface kinematic heat flux (w / m2)}

\item{d_0}{displacement height (m)}

\item{g}{acceleration due to gravity (9.81 m s -2 )}

\item{k}{von karman constant (0.4)}

\item{C_p}{1004.67 (specific heat at constant pressure J kg-1 K-1}
}
\value{
L in metres. stability (x) = z - d_0 / L
}
\description{
Used to determine atmospheric stability.
    Stable: x > 0
    Neutral: x ~= 0
    Unstable: x < 0
}
\details{
From shuttleworth et al., 2012 eq. 20.10
}
