% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trb_momentum.R
\name{optim_wind_params}
\alias{optim_wind_params}
\title{Least Squares Calculate Parameters Ustar, Z_0m, and, d_0 Given Measured Wind Speed and stability correction.}
\usage{
optim_wind_params(uMeas, zHeight, z_0m = NA, d_0 = NA, phi_m, k = 0.4)
}
\arguments{
\item{uMeas}{measured wind speed (m/s)}

\item{zHeight}{height of wind measurement (m)}

\item{d_0}{displacement height (m)}

\item{phi_m}{stability correction due to momentum. 0 is for neutral or stable case.}

\item{k}{0.4 von karmans constant}

\item{z0_m}{roughness length of momentum (m)}
}
\value{
List of optimization result including ustar (friction velocity), Z_0m (roughness length), d_0 (displacement height) Params are accessed by 'returnedobj'$...
}
\description{
Least Squares Calculate Parameters Ustar, Z_0m, and, d_0 Given Measured Wind Speed and stability correction.
}
\examples{
df <- data.frame(
uMeas = c(4.6, 6.0, 7.6, 9.0),

zHeight = c(1, 3, 10, 30)
)




}
