% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_corrections.R
\name{phi_unstble_hv}
\alias{phi_unstble_hv}
\title{UNSTABLE CASE: Stability Correction for Heat or Water Vapour}
\usage{
phi_unstble_hv(z_L)
}
\arguments{
\item{z_L}{}
}
\value{
stability correction to be used in the optim flux calculations, if z_l is less than zero returns NA.
}
\description{
UNSTABLE CASE: Stability Correction for Heat or Water Vapour
}
