% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trb_heat.R
\name{sensible_H_flux}
\alias{sensible_H_flux}
\title{Estimate Turbulent Sensible Heat Flux}
\usage{
sensible_H_flux(
  ref_temp,
  surf_temp,
  p_atm,
  rho_air,
  zHeight,
  u_star,
  z_0m,
  d_0 = NA,
  phi_h,
  k = 0.4,
  C_p = 1005
)
}
\arguments{
\item{ref_temp}{temperature at reference height}

\item{surf_temp}{temperautre of surface}

\item{p_atm}{atmospheric pressure kpa}

\item{rho_air}{density of moist air}

\item{zHeight}{height above the surface (metres)}

\item{u_star}{friction velocity}

\item{z_0m}{roughness length}

\item{d_0}{displacement height due to veg (m)}

\item{phi_h}{stability correction due to sensible heat. 0 is for neutral or stable case.}

\item{k}{(von Karman’s constant)}

\item{C_p}{(specific heat at constant pressure J kg-1 K-1}
}
\value{
J m-2 s-1 OR W m-2
}
\description{
Use this function if have all params listed below. If not use the optimiser function fit_neutral_temp. Taken from This formula is taken from the head and mass exchange lecture slide 19.
}
