% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_wv.R
\name{water_vapour_flux}
\alias{water_vapour_flux}
\title{Estimate Turbulent Water Vapour Flux}
\usage{
water_vapour_flux(
  ref_temp,
  surf_temp,
  p_atm,
  ref_rh,
  surf_rh,
  rho_air,
  zHeight,
  u_star,
  z_0m,
  d_0 = NA,
  k = 0.4,
  lambda = 2500000
)
}
\arguments{
\item{ref_temp}{temperature at reference height}

\item{surf_temp}{temperautre of surface}

\item{p_atm}{atmospheric pressure kpa}

\item{ref_rh}{relative humidity fraction at reference height}

\item{surf_rh}{relative humidity fraction at surface}

\item{rho_air}{density of moist air}

\item{zHeight}{height above the surface (metres)}

\item{u_star}{friction velocity}

\item{z_0m}{roughness length}

\item{d_0}{displacement height due to veg (m)}

\item{k}{0.4 (von Karman’s constant)}

\item{lambda}{2.5e-6 J kg-1 (latent head of vapourization)}
}
\value{
w / m2
}
\description{
Use this function if have all params listed below. If not use the optimiser function fit_neutral_wv. Taken from This formula is taken from the head and mass exchange lecture slide 19.
}
