% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_wind.R
\name{fit_neutral_wind}
\alias{fit_neutral_wind}
\title{Calculate Ustar, Z_0m, and, d_0 Given Measured Wind Speed}
\usage{
fit_neutral_wind(uMeas, zHeight, z_0m = NA, d_0 = NA, k = 0.4)
}
\arguments{
\item{uMeas}{measured wind speed (m/s)}

\item{zHeight}{height of wind measurement (m)}
}
\value{
List of optimization result. Params are accessed by 'returnedobj'$...
}
\description{
Calculate Ustar, Z_0m, and, d_0 Given Measured Wind Speed
}
\examples{
df <- data.frame(
uMeas = c(4.6, 6.0, 7.6, 9.0),

zHeight = c(1, 3, 10, 30)
)




}
