% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_neutral_wind.R
\name{friction_velocity}
\alias{friction_velocity}
\title{Calculate Friction Velocity}
\usage{
friction_velocity(uMeas, zHeight, d_0, z0_m, k = 0.4)
}
\arguments{
\item{uMeas}{measured wind speed m s-1}

\item{zHeight}{measurement height (m)}

\item{d_0}{displacement height (m)}

\item{z0_m}{roughness length of momentum (m)}

\item{k}{0.4 von karmans constant}
}
\value{
friction velocity (m/s)
}
\description{
Use this if you have all params listed below. If missing more than 1 use the fit_neitral_wind optimizer. This formula is taken from the head and mass exchange lecture slide 19
}
