% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trb_heat.R
\name{optim_heat_params}
\alias{optim_heat_params}
\title{Least Squares Estimate Turbulent Sensible Heat Flux Parameters}
\usage{
optim_heat_params(
  ref_temp,
  surf_temp,
  p_atm,
  rho_air,
  zHeight,
  u_star,
  z_0m,
  d_0 = NA,
  phi_h,
  k = 0.4,
  C_pl = 1004.67
)
}
\arguments{
\item{ref_temp}{temperature at reference height}

\item{surf_temp}{temperautre of surface}

\item{p_atm}{atmospheric pressure kpa}

\item{rho_air}{density of moist air}

\item{zHeight}{height above the surface (metres)}

\item{u_star}{friction velocity}

\item{z_0m}{roughness length}

\item{d_0}{displacement height due to veg (m)}

\item{phi_h}{stability correction due to sensible heat. 0 is for neutral or stable case.}

\item{k}{(von Karman’s constant)}
}
\value{

}
\description{
Least Squares Estimate Turbulent Sensible Heat Flux Parameters
}
